+++ tests/backend/lora/conversions/test_flux_lora_conversion_utils.py
@@ -3,7 +3,7 @@ import torch
 
 from invokeai.backend.flux.model import Flux
 from invokeai.backend.flux.util import params
-from invokeai.backend.lora.conversions.flux_lora_conversion_utils import (
+from invokeai.backend.lora.conversions.flux_kohya_lora_conversion_utils import (
     convert_flux_kohya_state_dict_to_invoke_format,
     is_state_dict_likely_in_flux_kohya_format,
     lora_model_from_flux_kohya_state_dict,
